Event

+++ src/main/java/edu/icet/repository/EventRepository.java
@@ -8,10 +8,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.sql.Date;
 import java.util.List;
 
-public interface EventRepository extends JpaRepository<EventEntity,Integer> {
+public interface EventRepository extends JpaRepository<EventEntity,Long> {
 
     List<EventEntity> findAllByEventType(EventType eventType);
     List<EventEntity> findAllByBudgetType(BudgetType budgetType);
     List<EventEntity> findAllByUserId(Long userId);
     List<EventEntity> findAllByEventDate(Date date);
 }
+
